autoremediatex-autoremediate-239075a211a6/pyproject.toml
[project]
name = "autoremediate"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "patchwork-cli @ file:///Users/pramilasingh/workspace/workspace_pradip_new/patchwork",
    "typer (>=0.15.2,<0.16.0)",
    "atlassian-python-api (>=4.0.3,<5.0.0)"
]

[tool.poetry]
packages = [{include = "autoremediate", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autoremediate = "autoremediate.cli:main"



